+++ packages/svelte2tsx/test/svelte2tsx/samples/module-snippet-component-instance-reference.v5/expectedv2.ts
@@ -1,0 +1,20 @@
+///<reference types="svelte" />
+;
+  // ...
+;;function $$render() {
+  const iconSnippet/*Ωignore_positionΩ*/ = ()/*Ωignore_startΩ*/: ReturnType<import('svelte').Snippet>/*Ωignore_endΩ*/ => { async ()/*Ωignore_positionΩ*/ => {
+   { const $$_nocI0C = __sveltets_2_ensureComponent(Icon); new $$_nocI0C({ target: __sveltets_2_any(), props: {  "size":16,}});}
+};return __sveltets_2_any(0)};
+  const/** @typedef {{ icon: any }} $$ComponentProps *//** @type {$$ComponentProps} */ { icon: Icon } = $props();
+;
+async () => {
+
+
+
+
+
+;__sveltets_2_ensureSnippet(iconSnippet());};
+return { props: /** @type {$$ComponentProps} */({}), exports: {}, bindings: __sveltets_$$bindings(''), slots: {}, events: {} }}
+const Input__SvelteComponent_ = __sveltets_2_fn_component($$render());
+/*Ωignore_startΩ*/type Input__SvelteComponent_ = ReturnType<typeof Input__SvelteComponent_>;
+/*Ωignore_endΩ*/export default Input__SvelteComponent_;